+++ packages/cheqd-blockchain-modules/src/did/internal.js
@@ -5,14 +5,20 @@ import {
 import { TypedUUID } from '@docknetwork/credential-sdk/types/generic';
 import { createInternalCheqdModule } from '../common';
 
-function parseDocument(document) {
+function createOrUpdateDidDocument(document) {
   return DIDDocument.from(document).toCheqd(this.types.DidDocument);
+}
+function deactivateDidDocument(id) {
+  return new CheqdDeactivateDidDocument(
+    this.types.Did.from(id),
+    TypedUUID.random(),
+  );
 }
 
 const methods = {
-  createDidDocument: parseDocument,
-  updateDidDocument: parseDocument,
-  deactivateDidDocument: (id) => new CheqdDeactivateDidDocument(id, TypedUUID.random()),
+  createDidDocument: createOrUpdateDidDocument,
+  updateDidDocument: createOrUpdateDidDocument,
+  deactivateDidDocument,
 };
 
 export class CheqdDIDModuleInternal extends createInternalCheqdModule(methods) {